3-DataUnify/RedshiftTransformation/requirements.txt
boto3
dbt-redshift
airflow-dbt
apache-airflow-providers-amazon